+++ tracing/jaeger/tracer.go
@@ -1,18 +1,19 @@
 package jaeger
 
 import (
+	"github.com/go-kit/kit/log"
 	"github.com/goph/fxt"
 	"github.com/opentracing/opentracing-go"
 	jaegercfg "github.com/uber/jaeger-client-go/config"
 	"github.com/uber/jaeger-lib/client/log/go-kit"
 )
 
-// NewTracer returns a new opentracing tracer.
+// NewTracer returns a new jaeger tracer.
 func NewTracer(params TracerParams) (opentracing.Tracer, error) {
 	var jaegerOptions []jaegercfg.Option
 
 	if params.Logger != nil {
-		jaegerOptions = append(jaegerOptions, jaegercfg.Logger(xkit.NewLogger(params.Logger)))
+		jaegerOptions = append(jaegerOptions, jaegercfg.Logger(xkit.NewLogger(log.With(params.Logger, "component", "jaeger"))))
 	}
 
 	if params.MetricsFactory != nil {